better

+++ Includes/libft.h
@@ -24,7 +24,7 @@ typedef struct s_list
 }				t_list;
 
 int				ft_is_valid(char *str);
-int				ft_atoi(const char *nptr);
+long			ft_atoi(const char *nptr);
 void			ft_bzero(void *s, size_t n);
 void			*ft_calloc(size_t count, size_t size);
 int				ft_isalnum(int c);
